requirements.txt
flask
werkzeug
ultralytics
openai-whisper
opencv-python-headless
google-generativeai
python-dotenv